refactor(agent): split registry.rs into submodules for code size limits
Split the 1105-line registry.rs (impl block 580 lines) into a directory
module with 4 files, each under 500 lines with impl blocks under 100
lines to comply with project code hard limits:

- registry/mod.rs (148 lines): types, constructors, SharedAgentRegistry
- registry/query.rs (279 lines): read-only query methods + tests
- registry/lifecycle.rs (257 lines): basic lifecycle methods + tests
- registry/cascade.rs (349 lines): cascade operations + tests

All 23 tests passing. SPEC.md section 3.1 updated accordingly.

Source: issue #213
Type: refactor
